Move set_service_units; remove unused imports
